Support new named networks correctly (#34)
Currently if the besu_network isn't one of the hard coded options, the value is ignored and a path to a custom genesis file is used. For the kiln network this results in the role failing because kiln is a built-in network and so besu_genesis_path is undefined.  This switches the logic to use the specified network, regardless of its value unless besu_genesis_path is specified so that new named networks are automatically supported but users can still specify their own genesis file.